backend/pyproject.toml
[project]
name = "rag-pdf-qna"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "mongoengine>=0.29.3",
    "openai>=2.30.0",
    "pinecone>=8.1.1",
    "pymongo[srv]>=4.16.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
